Create stock_picker function
The stock_picker function receives an array (stock_prices) that
represents a list of stock prices; it returns an array with the
indices of the best days to buy and sell.

To obtain that, first it creates an array to store all the possible
profits based on the stock_prices, then it makes an array that
calculates the profits. With those values now we can select the
highest one and finally we can obtain the indices of the best days to
buy and sell using the index method.